requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-genai
supabase
python-dotenv
pydantic[email]